+++ src/pywriter/model/__init__.py
@@ -7,9 +7,7 @@ chapter -- Provide a class for yWriter chapter representation.
 scene -- Provide a class for yWriter scene representation.
 world_element -- Provide a generic class for yWriter story world element representation.
 character -- Provide a class for yWriter character representation.
-cross_references -- Provide a class for yWriter cross reference generation.
 splitter -- Provide a helper class for scene and chapter splitting.
-id_generator -- Helper module for ID generation.
 
 Copyright (c) 2022 Peter Triesberger
 For further information see https://github.com/peter88213/PyWriter